chore: move plugin to different module
